+++ FurnitureRentalDomain/ReturnSummary.cs
@@ -5,7 +5,7 @@
         public int TotalCount { get; set; }
         public int OverdueCount { get; set; }
         public int EarlyCount { get; set; }
-        public double OverdueFine { get; set; }
-        public double EarlyRefund { get; set; }
+        public decimal OverdueFine { get; set; }
+        public decimal EarlyRefund { get; set; }
     }
 }
